requirements.txt
fastapi
uvicorn
openai
python-dotenv
jinja2
websockets
pymongo
requests
beautifulsoup4
langchain
langchain-community
langchain-cli
langchain-core
langchain-openai
langchain-groq